+++ fabric/src/main/java/whocraft/tardis_refined/common/capability/fabric/TardisLevelOperatorDummy.java
@@ -37,5 +37,10 @@ public class TardisLevelOperatorDummy extends TardisLevelOperatorImpl {
 
     }
 
+    @Override
+    public boolean deleteTARDIS() {
+        return false;
+    }
+
 
 }