+++ forge/src/main/java/whocraft/tardis_refined/common/dimension/forge/DimensionHandlerImpl.java
@@ -1,110 +1,24 @@
 package whocraft.tardis_refined.common.dimension.forge;
 
-import com.google.common.collect.ImmutableList;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
-import com.mojang.serialization.Lifecycle;
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.MappedRegistry;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.Registries;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.level.ServerChunkCache;
+import net.minecraft.core.*;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.server.level.TicketType;
-import net.minecraft.server.level.progress.ChunkProgressListener;
-import net.minecraft.util.Unit;
-import net.minecraft.world.RandomSequences;
-import net.minecraft.world.level.ChunkPos;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.biome.BiomeManager;
-import net.minecraft.world.level.border.BorderChangeListener;
-import net.minecraft.world.level.dimension.LevelStem;
-import net.minecraft.world.level.storage.DerivedLevelData;
-import net.minecraft.world.level.storage.LevelStorageSource;
-import net.minecraft.world.level.storage.WorldData;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.level.LevelEvent;
-import whocraft.tardis_refined.common.dimension.DimensionHandler;
-import whocraft.tardis_refined.common.network.messages.sync.S2CSyncLevelList;
-
-import java.util.concurrent.Executor;
-import java.util.function.BiFunction;
-
-import static whocraft.tardis_refined.common.util.forge.PlatformImpl.getServer;
 
 public class DimensionHandlerImpl {
 
+    public static <T> void unfreezeRegistry(MappedRegistry<T> registry) {
+        registry.unfreeze(); // This Neoforge method is deprecated so we may need to use an Accessor Mixin in the future.
+    }
 
-    public static ServerLevel createDimension(Level level, ResourceKey<Level> id) {
+    public static void onDimensionLoaded(ServerLevel level) {
+        level.getServer().markWorldsDirty();
+        MinecraftForge.EVENT_BUS.post(new LevelEvent.Load(level));
+    }
 
-        BiFunction<MinecraftServer, ResourceKey<LevelStem>, LevelStem> dimensionFactory = DimensionHandler::formLevelStem;
-
-        MinecraftServer server = getServer();
-        ServerLevel overworld = server.getLevel(Level.OVERWORLD);
-
-        final ResourceKey<LevelStem> dimensionKey = ResourceKey.create(Registries.LEVEL_STEM, id.location());
-
-        LevelStem dimension = dimensionFactory.apply(server, dimensionKey);
-
-        ChunkProgressListener chunkListener = server.progressListenerFactory.create(11);
-        Executor executor = server.executor;
-        LevelStorageSource.LevelStorageAccess levelSave = server.storageSource;
-
-
-        WorldData serverConfig = server.getWorldData();
-        DerivedLevelData derivedWorldInfo = new DerivedLevelData(serverConfig, serverConfig.overworldData());
-
-        //Actually register our dimension
-        Registry<LevelStem> dimensionRegistry = server.registryAccess().registryOrThrow(Registries.LEVEL_STEM);
-        if (dimensionRegistry instanceof MappedRegistry<LevelStem> writableRegistry) {
-            writableRegistry.unfreeze(); //Must unfreeze registry to allow our dimension to persist. This Neoforge method is deprecated so we may need to use an Accessor Mixin in the future.
-            writableRegistry.register(dimensionKey, dimension, Lifecycle.stable());
-        }
-        else {
-            throw new IllegalStateException(String.format("Unable to register dimension %s -- dimension registry not writable", dimensionKey.location()));
-        }
-
-        // now we have everything we need to create the world instance
-        ServerLevel newLevel = new ServerLevel(
-                server,
-                executor,
-                levelSave,
-                derivedWorldInfo,
-                id,
-                dimension,
-                chunkListener,
-                false, // boolean: is-debug-world
-                BiomeManager.obfuscateSeed(serverConfig.worldGenOptions().seed()),
-                ImmutableList.of(), // "special spawn list"
-                // phantoms, raiders, travelling traders, cats are overworld special spawns
-                // the dimension loader is hardcoded to initialize preexisting non-overworld worlds with no special spawn lists
-                // so this can probably be left empty for best results and spawns should be handled via other means
-                false, // "tick time", true for overworld, always false for everything else
-                new RandomSequences(BiomeManager.obfuscateSeed(serverConfig.worldGenOptions().seed())));
-
-        DimensionHandler.addDimension(newLevel.dimension());
-
-        overworld.getWorldBorder().addListener(new BorderChangeListener.DelegateBorderChangeListener(newLevel.getWorldBorder()));
-
-        server.levels.put(id, newLevel);
-
-        server.markWorldsDirty();
-
-        MinecraftForge.EVENT_BUS.post(new LevelEvent.Load(newLevel));
-
-        new S2CSyncLevelList(newLevel.dimension(), true).sendToAll();
-
-        BlockPos blockPos = new BlockPos(0, 0, 0);
-        ChunkPos chunkPos = new ChunkPos(blockPos);
-        chunkListener.updateSpawnPos(chunkPos);
-        ServerChunkCache serverchunkcache = newLevel.getChunkSource();
-        serverchunkcache.getLightEngine().checkBlock(blockPos); //Runs lighting update
-        serverchunkcache.addRegionTicket(TicketType.START, chunkPos, 11, Unit.INSTANCE);
-        server.tell(server.wrapRunnable(chunkListener::stop));
-
-        return newLevel;
+    public static void onDimensionUnloaded(ServerLevel level) {
+        level.getServer().markWorldsDirty();
+        MinecraftForge.EVENT_BUS.post(new LevelEvent.Unload(level));
     }
 
 }